Add big number + reorder table on register homepage.
Add a big number at the top to show the total number of entries
in the register. Style this responsively. Eventually it would
be nice to have more metadata here (e.g. last updated).

Add section headings on the page to differentiate the introduction
and metadata from the register itself.

Reorder the columns in the table to be more intuitive. Name goes
first, then all the attribute columns in alphabetical order, and
finally the hash - which we rename to 'last update'. This still
needs work to make the distinction between name and hash easier
to understand.

Ensure that the table is scrollable if it overflows. We may want
to hack in a scrollbar at the top of the table too (for JS users)
so it's more visible.