Add simple search and filter to register table.
Use the pure JS library list.js [1] to allow simple searching
and filtering on register tables. This allows users to find
items of interest quickly and get a sense of the data quickly.

Add some lightweight frontend styling for the search and sort
options.

This may not perform well on very large datasets - for these
we may want to look into a more powerful library like SlickGrid
or we may prefer simply to offer a subset of the data, and make
this clear in the UI.

We will also need to look into how it interacts with pagination.
I'll do this once the education data is loaded.

[1] http://www.listjs.com/