Fix build
Specifically pin the version of `openregister-python` (previously
`openregister-entry`) that we require. I've chosen this version fairly
arbitrarily as being a version that was released at a similar time to
when `openregister-flask` was last worked on.

It appears as though the method signatures have changed significantly
enough that using the latest version of `openregister-python` is not
simple. At the very least this change makes the build green.